decrating: absorb shipper-process into shipper::ops::process
Move shipper-process crate (1948 LOC) into crates/shipper/src/ops/process/
as crate-private modules, split into types.rs, run.rs, timeout.rs, which.rs,
cargo.rs + mod.rs facade, tests.rs, snapshot_tests.rs, and
cross_platform_edge_case_tests.rs. Delete the stale 105-LOC duplicate at
crates/shipper/src/process.rs and the 32-LOC shim at process_micro.rs.
Remove shipper-process from workspace members and from shipper's deps.
Drop the micro-process feature flag (from both shipper and shipper-cli
Cargo.toml). Add `which = "8.0"` as a shipper dep and enable `yaml`
feature on insta dev-dep to support the co-located snapshot tests.

Per docs/decrating-plan.md §6 Phase 2.
