+++ crates/shipper-cli/Cargo.toml
@@ -40,7 +40,6 @@ micro-storage = ["shipper/micro-storage"]
 micro-cargo = ["shipper/micro-cargo"]
 micro-plan = ["shipper/micro-plan"]
 micro-registry = ["shipper/micro-registry"]
-micro-process = ["shipper/micro-process"]
 micro-webhook = ["shipper/micro-webhook"]
 micro-types = ["shipper/micro-types"]
 micro-config = ["shipper/micro-config"]
