--- crates/shipper-process/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "shipper-process"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-rust-version.workspace = true
-description = "Process execution for shipper"
-repository = "https://github.com/EffortlessMetrics/shipper"
-documentation = "https://docs.rs/shipper-process"
-keywords = ["process", "command", "execution"]
-categories = ["development-tools"]
-
-[dependencies]
-anyhow = "1.0"
-serde.workspace = true
-which = "8.0"
-humantime = "2"
-
-[dev-dependencies]
-tempfile = "3"
-serde_json = "1.0"
-proptest.workspace = true
-insta = { workspace = true, features = ["yaml"] }
-temp-env = "0.3"
-
-[lints]
-workspace = true
