+++ crates/shipper/Cargo.toml
@@ -26,7 +26,6 @@ shipper-cargo = { path = "../shipper-cargo", optional = true }
 shipper-engine-parallel = { path = "../shipper-engine-parallel", optional = true }
 shipper-plan = { path = "../shipper-plan", optional = true }
 shipper-registry = { path = "../shipper-registry", optional = true }
-shipper-process = { path = "../shipper-process", optional = true }
 shipper-webhook = { path = "../shipper-webhook" }
 shipper-types = { path = "../shipper-types" }
 shipper-config = { path = "../shipper-config" }
@@ -50,6 +49,7 @@ reqwest = { version = "0.13.2", features = ["blocking", "json", "rustls"] }
 rand = { version = "0.10.1", features = ["std"] }
 chrono = { version = "0.4.44", features = ["serde"] }
 humantime = "2.3.0"
+which = "8.0"
 toml = "1.0.3"
 hex = "0.4.3"
 gethostname = "1.1.0"
@@ -72,14 +72,13 @@ micro-storage = ["dep:shipper-storage"]
 micro-cargo = ["dep:shipper-cargo"]
 micro-plan = ["dep:shipper-plan", "micro-types"]
 micro-registry = ["dep:shipper-registry"]
-micro-process = ["dep:shipper-process"]
 micro-webhook = []
 micro-types = ["micro-webhook", "micro-encrypt", "micro-config"]
 micro-config = ["micro-types", "micro-storage"]
 micro-state = ["dep:shipper-state", "micro-types", "micro-environment", "micro-encrypt"]
 micro-store = ["dep:shipper-store", "micro-types", "micro-events", "micro-state"]
 micro-parallel = ["micro-types", "micro-plan", "micro-registry", "dep:shipper-engine-parallel"]
-micro-all = ["micro-auth", "micro-git", "micro-events", "micro-encrypt", "micro-environment", "micro-storage", "micro-cargo", "micro-plan", "micro-registry", "micro-process", "micro-webhook", "micro-types", "micro-config", "micro-state", "micro-store", "micro-parallel"]
+micro-all = ["micro-auth", "micro-git", "micro-events", "micro-encrypt", "micro-environment", "micro-storage", "micro-cargo", "micro-plan", "micro-registry", "micro-webhook", "micro-types", "micro-config", "micro-state", "micro-store", "micro-parallel"]
 
 [lints]
 workspace = true
